Allow Import of .csv and other dsv data files
Fixes #109. As a hotfix, modifies the behaviour of pspm_get_text to allow for the use of non Matlab standard column separators in delimiter separated values files. 

These modifications should not break backwards compatibility, but this has to be verified before rolling it out. A cleaner implementation is discussed in #110